+++ src/CompilerCxx/Naming.hs
@@ -146,10 +146,10 @@ valueCustom n = "ExtValue_" ++ show n
 callName :: FunctionName -> String
 callName f = "Call_" ++ show f
 
-functionDebugName :: ScopedFunction c -> String
-functionDebugName f
-  | sfScope f == CategoryScope = show (sfType f) ++ ":" ++ show (sfName f)
-  | otherwise                  = show (sfType f) ++ "." ++ show (sfName f)
+functionDebugName :: CategoryName -> ScopedFunction c -> String
+functionDebugName t f
+  | sfScope f == CategoryScope = show t ++ ":" ++ show (sfName f)
+  | otherwise                  = show t ++ "." ++ show (sfName f)
 
 functionName :: ScopedFunction c -> String
 functionName f = "Function_" ++ show (sfType f) ++ "_" ++ show (sfName f)